small bug fixes && changes
